+++ src/SFA.DAS.Payments.AcceptanceTests.Core/Data/LearnerEarningsHistory.cs
@@ -5,7 +5,7 @@ namespace SFA.DAS.Payments.AcceptanceTests.Core.Data
 {
     public class LearnerEarningsHistory
     {
-        public IEnumerable<AdditionalIlrData> AdditionalData { get; set; }
+        public AdditionalIlrData AdditionalData { get; set; }
 
         public IEnumerable<Earning> PreviousEarnings { get; set; }
     }
